fix(exceptions): move MissingResourceError to __init__.py to resolve circular import
Define DiagnosticContext and MissingResourceError directly in
__init__.py after UserError, and keep diagnostic_errors.py as
a re-export module for backward compatibility.
